fix: issue with headers transformed by node.js
Adding kassette in a scenario can sometimes be a problem because of how
Node.js transforms headers.
A problem can happen especially with the WWW-Authenticate header whose
duplicate values must not be merged together into a single comma-separated
value as Node.js does.
This commit uses rawHeaders instead of headers to fix this kind of problems,
so that we preserve as much as possible the case of headers and if they are
duplicated across multiple lines, while still trying to keep as much
as possible compatibility with existing usages of kassette.