fix(test): e2e code coverage on windows
On Windows, it looks like readFileSync is called with a path containing
forward slashes (instead of the usual backward slashes), and
shouldInstrument checks that the file path starts with the correct
folder (with backward slashes), so no file was instrumented.
Calling normalize replaces forward slashes with backward slashes and
fixes the issue.