fix: allow connection to unsafe legacy servers
This commit adds the SSL_OP_LEGACY_SERVER_CONNECT flag when connecting to
legacy remote servers for increased compatibility.
Note that it allows man-in-the-middle attacks, but those attacks are
already possible anyway because of the rejectUnauthorized option which
allows any certificate on the server. kassette is a test tool that is not
designed to run in a production environment.

This PR fixes the following error that can happen when this flag is not
set:

78250000:error:0A000152:SSL routines:final_renegotiate:unsafe legacy renegotiation disabled